4.21 valmis

+++ osa4/blogilista/controllers/blogs.js
@@ -33,8 +33,24 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+
+  if (blog && user && blog.user.toString() === user._id.toString()) {
+    user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id)
+    await user.save()
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'blog is not made by this account' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
